+++ tests/testthat/test-to_labelled.r
@@ -36,3 +36,44 @@ test_that("to_labelled.factor accepts non continuous labels", {
     labelled(c(1, 1, 2, 2, 9, 2, 1, 9), c("yes" = 1, "no" = 2, "don't know" = 9))
   )
 })
+
+test_that("to_labelled.factor works with '[code] label' factors", {
+  l <- labelled(c(1, 1, 2, 2, 9, 2, 1, 9), c("yes" = 1, "no" = 2, "don't know" = 9))
+  expect_equal(
+    to_factor(l, levels = "p") %>% to_labelled(),
+    l
+  )
+
+  l <- labelled(
+    c("M", "M", "F", "X", "N/A"),
+    c(Male = "M", Female = "F", Refused = "X", "Not applicable" = "N/A")
+  )
+  expect_equal(
+    to_factor(l, levels = "p") %>% to_labelled(),
+    l
+  )
+
+  # if labels is provided apply normal rule
+  l <- labelled(c(1, 1, 2, 2, 9, 2, 1, 9), c("yes" = 1, "no" = 2, "don't know" = 9))
+  f <- to_factor(l, levels = "p")
+  x <- f %>% to_labelled(labels = c("[1] yes" = 123, "[2] no" = 456))
+  expect_equivalent(
+    unclass(x),
+    c(123, 123, 456, 456, NA, 456, 123, NA)
+  )
+
+  # should not be applied if duplicates in code
+  f <- factor(c("[1] yes", "[2] no", "[1] don't know"))
+  expect_warning(l <- to_labelled(f))
+  expect_warning(l <- to_labelled(f, .quiet = TRUE), NA)
+  expect_identical(
+    names(val_labels(l)),
+    levels(f)
+  )
+
+  # check potential duplicates in numerical codes
+  f <- factor(c("[1] yes", "[1.0] no", "[01] don't know"))
+  expect_warning(to_labelled(f))
+  expect_warning(to_labelled(f, .quiet = TRUE), NA)
+  expect_true(is.character(to_labelled(f, .quiet = TRUE)))
+})